fix(server): fixes after cr

+++ apps/server/src/modules/user/user.service.ts
@@ -1,8 +1,13 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
 import { User } from './entity/user.entity';
 import { UserAlreadyExistsException } from './exceptions';
 import * as bcrypt from 'bcrypt';
+import { plainToInstance } from 'class-transformer';
+import {
+  UserResponseDto,
+  UserResponseSchema,
+} from '@nonogram-api-monorepo/types';
 
 @Injectable()
 export class UserService {
@@ -27,37 +32,65 @@ export class UserService {
       password: await bcrypt.hash(createUserDto.password, salt),
     };
 
-    this.userModel.create(createUserDto);
+    return UserResponseSchema.parse(
+      (await this.userModel.create(createUserDto)).toJSON()
+    );
   }
 
-  async findOne(personalNumber): Promise<User | null> {
-    return this.userModel.findOne({
+  async getUserByPersonalNumber(personalNumber): Promise<User | null> {
+    return await this.userModel.findOne({
       where: {
         personalNumber: personalNumber,
       },
     });
   }
 
-  async getUser(id) {
-    return this.userModel.findOne({
-      where: { id },
-    });
+  async getUserById(paramId, userId) {
+    if (paramId.id !== userId) {
+      throw new UnauthorizedException();
+    }
+
+    return UserResponseSchema.parse(
+      (
+        await this.userModel.findOne({
+          where: { id: userId },
+        })
+      ).toJSON()
+    );
   }
 
   async updateUser(id, userUpdateDto) {
+    if (id !== userUpdateDto.id) {
+      throw new UnauthorizedException();
+    }
+
     const user = await this.userModel.findOne({
       where: { id },
     });
+
+    if (userUpdateDto.password) {
+      const saltRounds = parseInt(process.env.BCRYPT_SALT);
+      const salt = await bcrypt.genSalt(saltRounds);
+
+      userUpdateDto = {
+        ...userUpdateDto,
+        password: await bcrypt.hash(userUpdateDto.password, salt),
+      };
+    }
 
     user.set({
       ...userUpdateDto,
     });
 
-    await user.save();
+    return UserResponseSchema.parse((await user.save()).toJSON());
   }
 
-  deleteUser(id) {
-    this.userModel.destroy({
+  async deleteUser(id, userId) {
+    if (id !== userId) {
+      throw new UnauthorizedException();
+    }
+
+    return await this.userModel.destroy({
       where: { id },
     });
   }
